rename puppet folder name to `abstract-puppet`
