+++ src/abstract-puppet/index.ts
