+++ src/abstract-puppet/message.ts
