+++ src/abstract-puppet/puppet-accessory.spec.ts
