+++ src/abstract-puppet/puppet-accessory.ts
