+++ src/abstract-puppet/puppet.spec.ts
