+++ src/abstract-puppet/puppet.ts
