+++ src/abstract-puppet/room.ts
