+++ src/puppet-mock/mock-friend-request.ts
@@ -21,7 +21,7 @@ import {
   // config,
   log,
 }                     from '../config'
-import FriendRequest  from '../puppet/friend-request'
+import FriendRequest  from '../abstract-puppet/friend-request'
 
 import MockContact     from './mock-contact'
